grpc-test-vkr-main/vkr-glossary-grpc-project/glossary-grpc/glossary-service/requirements.txt
grpcio ~= 1.30
grpcio-tools ~= 1.30

